Add phpunit action for running code coverage.
This action includes PCOV and will enable consumers to run PHPUnit 8
using PHP 7.2. Both are the officially supported versions for
phpDocumentor and its libraries.

It is _not_ recommended to use this image to test against operating
systems other than the included one; as you won't notice anything
breaking in windows or mac as this is, of course, a Docker image.